LayoutTests/model-element/immersive/model-element-immersive-iframe-restrictions-expected.txt


PASS document.immersiveEnabled is true on top-level document
PASS document.immersiveEnabled is false on iframe document
PASS requestImmersive rejects with InvalidAccessError in iframe
PASS requestImmersive error message is correct for iframe
PASS immersiveerror event fires for iframe restriction
